rename server status to server state view
